% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{calculate_running_std}
\alias{calculate_running_std}
\title{An internal method to calculate running standard deviation from the scaled
running variance.}
\usage{
calculate_running_std(this)
}
\description{
An internal method to calculate running standard deviation from the scaled
running variance.
}
