% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{combine_pair}
\alias{combine_pair}
\title{Combines two Hermite estimators}
\usage{
combine_pair(this, hermite_estimator_other)
}
\arguments{
\item{this}{A hermite_estimator object. The first Hermite series based 
estimator.}

\item{hermite_estimator_other}{A hermite_estimator object. The second Hermite
series based estimator.}
}
\value{
An object of class hermite_estimator.
}
\description{
This method allows a pair of Hermite based estimators of class 
hermite_estimator to be consistently combined.
}
\details{
Note that the N and standardize arguments must be the same for the two 
estimators in order to combine them. In addition, note that exponentially
weighted estimators cannot be combined.
}
