% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{combine_standardized_helper}
\alias{combine_standardized_helper}
\title{Internal method to combine a list of standardized Hermite estimators with 
improved accuracy}
\usage{
combine_standardized_helper(hermite_estimators)
}
\arguments{
\item{hermite_estimators}{A list of hermite_estimator objects.}
}
\value{
An object of class hermite_estimator.
}
\description{
Internal method to combine a list of standardized Hermite estimators with 
improved accuracy
}
