% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_bivar.R
\name{combine_standardized_helper_bivar}
\alias{combine_standardized_helper_bivar}
\title{Internal method to combine a list of standardized bivariate Hermite 
estimators with improved accuracy}
\usage{
combine_standardized_helper_bivar(hermite_estimators)
}
\arguments{
\item{hermite_estimators}{A list of hermite_estimator_bivar objects.}
}
\value{
An object of class hermite_estimator_bivar.
}
\description{
Internal method to combine a list of standardized bivariate Hermite 
estimators with improved accuracy
}
