% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{combine_standardized_helper_univar}
\alias{combine_standardized_helper_univar}
\title{Internal method to combine a list of standardized Hermite estimators with 
improved accuracy}
\usage{
combine_standardized_helper_univar(hermite_estimators)
}
\arguments{
\item{hermite_estimators}{A list of hermite_estimator_univar objects.}
}
\value{
An object of class hermite_estimator_univar.
}
\description{
Internal method to combine a list of standardized Hermite estimators with 
improved accuracy
}
