% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{cum_prob.hermite_estimator_univar}
\alias{cum_prob.hermite_estimator_univar}
\title{Estimates the cumulative probability for a vector of x values}
\usage{
\method{cum_prob}{hermite_estimator_univar}(h_est_obj, x, clipped = FALSE, accelerate_series = TRUE)
}
\arguments{
\item{h_est_obj}{A hermite_estimator_univar object.}

\item{x}{A numeric vector. Values at which to estimate the cumulative
probability}

\item{clipped}{A boolean value. This value determines whether cumulative
probabilities are clipped to lie within the range [0,1].}

\item{accelerate_series}{A boolean value. This value determines whether
Hermite series acceleration is applied.}
}
\value{
A numeric vector of cumulative probability values.
}
\description{
This method calculates the cumulative probability values at a vector of
x values using the hermite_estimator_univar object (h_est_obj).
}
\details{
The object must be updated with observations prior to the use of this method.
}
