% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{cum_prob_quantile_helper}
\alias{cum_prob_quantile_helper}
\title{Estimates the cumulative probability for quantile estimation}
\usage{
cum_prob_quantile_helper(this, x)
}
\arguments{
\item{this}{A hermite_estimator object.}

\item{x}{A numeric vector.}
}
\value{
A numeric vector of cumulative probability values.
}
\description{
This helper method uses a modified distribution function estimator which
differs from the cum_prob.hermite_estimator.
}
\details{
The modified distribution function estimator appears more accurate for
quantile estimation as validated empirically in:

\url{https://projecteuclid.org/euclid.ejs/1488531636}

This method is intended for internal use by the hermite_estimator class.
}
