% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{dens}
\alias{dens}
\title{Estimates the probability density for a vector of x values}
\usage{
dens(this, x, clipped)
}
\arguments{
\item{this}{A hermite_estimator object.}

\item{x}{A numeric vector. Values at which to estimate the probability 
density.}

\item{clipped}{A boolean value. This value determines whether 
probability densities are clipped to be bigger than zero.}
}
\value{
A numeric vector of probability density values.
}
\description{
This method calculates the probability density values at a vector of 
x values using the hermite_estimator object (this).
}
\details{
The object must be updated with observations prior to the use of the method.
}
