% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{describe_estimator}
\alias{describe_estimator}
\title{Internal helper function for print methods.}
\usage{
describe_estimator(this, est_type)
}
\description{
Internal helper function for print methods.
}
