% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gauss_hermite_quad_100}
\alias{gauss_hermite_quad_100}
\title{Calculates \eqn{\int_{-\infty}^{\infty} f(x) e^{-x^2} dx} using 
Gauss-Hermite quadrature with 100 terms.}
\usage{
gauss_hermite_quad_100(f)
}
\arguments{
\item{f}{A function.}
}
\value{
A numeric value.
}
\description{
Calculates \eqn{\int_{-\infty}^{\infty} f(x) e^{-x^2} dx} using 
Gauss-Hermite quadrature with 100 terms.
}
