% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{get_coefficients}
\alias{get_coefficients}
\title{Returns Hermite series expansion coefficients}
\usage{
get_coefficients(this)
}
\arguments{
\item{this}{A hermite_estimator object.}
}
\value{
Numeric vector of length N+1
}
\description{
Returns Hermite series expansion coefficients
}
\examples{
hermite_est_1 <- hermite_estimator(N = 10, standardize = FALSE)
hermite_est_1 <- update_batch(hermite_est_1, rnorm(30))
coefficient_vec <- get_coefficients(hermite_est_1)
}
