% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hermite_function}
\alias{hermite_function}
\title{Outputs orthonormal Hermite functions}
\usage{
hermite_function(N, x, normalization)
}
\arguments{
\item{N}{An integer number.}

\item{x}{A numeric vector.}

\item{normalization}{A numeric vector of normalization factors generated by
the hermite_normalization function.}
}
\value{
A numeric matrix with N+1 rows and length(x) columns.
}
\description{
The method calculates the orthonormal Hermite functions, \eqn{h_k(x)} 
from \eqn{k=0,\dots,N} for the vector of values, x.
}
\author{
Michael Stephanou <michael.stephanou@gmail.com>
}
