% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hermite_function_sum}
\alias{hermite_function_sum}
\title{Outputs the sum of orthonormal Hermite functions}
\usage{
hermite_function_sum(N, x, normalization)
}
\arguments{
\item{N}{An integer number.}

\item{x}{A numeric vector.}

\item{normalization}{A numeric vector of normalization factors generated by
the hermite_normalization function.}
}
\value{
A numeric vector of length N+1.
}
\description{
The method calculates the sum of orthonormal Hermite functions, 
\eqn{\sum_{i} h_k(x_{i})} from \eqn{k=0,\dots,N} for the vector of values,
x.
}
\author{
Michael Stephanou <michael.stephanou@gmail.com>
}
