% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hermite_int_full}
\alias{hermite_int_full}
\title{Convenience function to output the integral of the orthonormal Hermite 
functions on the full domain}
\usage{
hermite_int_full(N)
}
\arguments{
\item{N}{An integer number.}
}
\value{
A numeric matrix with N+1 rows and 1 columns.
}
\description{
The method calculates \eqn{\int_{-\infty}^{\infty} h_k(t) dt} 
for \eqn{k=0,\dots,N}.
}
