% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hermite_int_full_domain}
\alias{hermite_int_full_domain}
\title{Outputs integral of the orthonormal Hermite functions on the full domain}
\usage{
hermite_int_full_domain(N)
}
\arguments{
\item{N}{An integer number.}
}
\value{
A numeric matrix with N+1 rows and 1 columns.
}
\description{
The method calculates \eqn{\int_{-\infty}^{\infty} h_k(t) dt} 
for \eqn{k=0,\dots,N}.
}
\author{
Michael Stephanou <michael.stephanou@gmail.com>
}
