% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hermite_polynomial}
\alias{hermite_polynomial}
\title{Outputs physicist version of Hermite Polynomials}
\usage{
hermite_polynomial(N, x)
}
\arguments{
\item{N}{An integer number.}

\item{x}{A numeric vector.}
}
\value{
A numeric matrix with N+1 rows and length(x) columns.
}
\description{
The method calculates the physicist version of Hermite polynomials, 
\eqn{H_k(x)} from \eqn{k=0,\dots,N} for the vector of values, x.
}
\author{
Michael Stephanou <michael.stephanou@gmail.com>
}
