% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hermite_polynomial_N}
\alias{hermite_polynomial_N}
\title{Convenience function to output physicist Hermite polynomials
 
 
The method calculates the physicist version of Hermite polynomials, 
\eqn{H_k(x)} from \eqn{k=0,\dots,N} for the vector of values, x.}
\usage{
hermite_polynomial_N(N, x)
}
\arguments{
\item{N}{An integer number.}

\item{x}{A numeric vector.}
}
\value{
A numeric matrix with N+1 rows and length(x) columns.
}
\description{
Convenience function to output physicist Hermite polynomials
 
 
The method calculates the physicist version of Hermite polynomials, 
\eqn{H_k(x)} from \eqn{k=0,\dots,N} for the vector of values, x.
}
