\name{hermiter-package}
\alias{hermiter-package}
\alias{hermiter}
\docType{package}
\title{\packageTitle{hermiter}}
\description{\packageDescription{hermiter}}
\section{Package Content}{\packageIndices{hermiter}}
\author{\packageAuthor{hermiter}}
\section{Maintainer}{\packageMaintainer{hermiter}}
\keyword{package}
