% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_bivar.R
\name{initialize_batch_bivar}
\alias{initialize_batch_bivar}
\title{Initializes the Hermite series based estimator with a batch of data}
\usage{
initialize_batch_bivar(h_est_obj, x)
}
\arguments{
\item{h_est_obj}{A hermite_estimator_bivar object.}

\item{x}{A numeric matrix. A matrix of bivariate observations to be 
incorporated into the estimator. Each row corresponds to a single bivariate
observation.}
}
\value{
An object of class hermite_estimator_bivar.
}
\description{
Initializes the Hermite series based estimator with a batch of data
}
