% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{initialize_batch_univar}
\alias{initialize_batch_univar}
\title{Initializes the Hermite series based estimator with a batch of data}
\usage{
initialize_batch_univar(h_est_obj, x)
}
\arguments{
\item{h_est_obj}{A hermite_estimator_univar object.}

\item{x}{A numeric vector. A vector of observations to be incorporated
into the estimator.}
}
\value{
An object of class hermite_estimator_univar.
}
\description{
Initializes the Hermite series based estimator with a batch of data
}
