% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{IQR.default}
\alias{IQR.default}
\title{Estimates the Interquartile range (IQR)}
\usage{
\method{IQR}{default}(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Optional additional arguments to the stats::IQR function.}
}
\value{
A numeric value.
}
\description{
This creates a default generic method for the stats::IQR function.
}
