% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{IQR.hermite_estimator_univar}
\alias{IQR.hermite_estimator_univar}
\title{Estimates the Interquartile range (IQR)}
\usage{
\method{IQR}{hermite_estimator_univar}(x, ...)
}
\arguments{
\item{x}{A hermite_estimator_univar object.}

\item{...}{Optional additional arguments to the quant function namely 
algorithm and accelerate_series.}
}
\value{
A numeric value.
}
\description{
This generic method is a convenience wrapper around the quant method to 
calculate the interquartile range.
}
