% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{IQR}
\alias{IQR}
\title{Estimates the Interquartile range (IQR)}
\usage{
IQR(x, ...)
}
\arguments{
\item{x}{A numeric vector or hermite_estimator_univar object.}

\item{...}{Optional additional arguments.}
}
\value{
A numeric value.
}
\description{
This generic method dispatches to the stats::IQR function or the 
IQR.hermite_estimator_univar function depending on the class of x.
}
