% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{merge_moments_and_count_univar}
\alias{merge_moments_and_count_univar}
\title{Internal method to consistently merge the number of observations, means and 
variances of two Hermite estimators}
\usage{
merge_moments_and_count_univar(hermite_estimator1, hermite_estimator2)
}
\arguments{
\item{hermite_estimator1}{A hermite_estimator_univar object.}

\item{hermite_estimator2}{A hermite_estimator_univar object.}
}
\value{
An object of class hermite_estimator_univar.
}
\description{
The algorithm to merge the variances consistently comes from
Schubert, Erich, and Michael Gertz. "Numerically stable parallel computation 
of (co-) variance." Proceedings of the 30th International Conference on 
Scientific and Statistical Database Management. 2018.
}
