% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{merge_standardized_helper_univar}
\alias{merge_standardized_helper_univar}
\title{Internal method to merge a list of standardized Hermite estimators}
\usage{
merge_standardized_helper_univar(hermite_estimators)
}
\arguments{
\item{hermite_estimators}{A list of hermite_estimator_univar objects.}
}
\value{
An object of class hermite_estimator_univar.
}
\description{
Internal method to merge a list of standardized Hermite estimators
}
