% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{plot.hcdf_bivar}
\alias{plot.hcdf_bivar}
\title{Plots the hcdf_bivar object as output by the hcdf function when evaluated
on a hermite_estimator_bivar object.}
\usage{
\method{plot}{hcdf_bivar}(x, main = "Hermite CDF", xlab = "X", ylab = "Y", ...)
}
\arguments{
\item{x}{A hcdf_bivar object.}

\item{main}{A string, title for plot.}

\item{xlab}{A string, x label for plot.}

\item{ylab}{A string, y label for plot.}

\item{...}{Unused.}
}
\description{
Plots the hcdf_bivar object as output by the hcdf function when evaluated
on a hermite_estimator_bivar object.
}
