% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{plot.hcdf_univar}
\alias{plot.hcdf_univar}
\title{Plots the hcdf_univar object as output by the hcdf function when evaluated
on a hermite_estimator_univar object.}
\usage{
\method{plot}{hcdf_univar}(x, main = "Hermite CDF", xlab = "x", ylab = "F(x)", ...)
}
\arguments{
\item{x}{A hcdf_univar object.}

\item{main}{A string, title for plot.}

\item{xlab}{A string, x label for plot.}

\item{ylab}{A string, y label for plot.}

\item{...}{Additional parameters for plotting.}
}
\description{
Plots the hcdf_univar object as output by the hcdf function when evaluated
on a hermite_estimator_univar object.
}
