% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{plot.hdensity_bivar}
\alias{plot.hdensity_bivar}
\title{Plots the hdensity_bivar object as output by the density function when 
evaluated on a hermite_estimator_bivar object.}
\usage{
\method{plot}{hdensity_bivar}(x, main = "Hermite PDF", xlab = "X", ylab = "Y", ...)
}
\arguments{
\item{x}{A hdensity_bivar object.}

\item{main}{A string, title for plot.}

\item{xlab}{A string, x label for plot.}

\item{ylab}{A string, y label for plot.}

\item{...}{Unused.}
}
\description{
Plots the hdensity_bivar object as output by the density function when 
evaluated on a hermite_estimator_bivar object.
}
