% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{plot.hdensity_univar}
\alias{plot.hdensity_univar}
\title{Plots the hdensity_univar object as output by the density function when 
evaluated on a hermite_estimator_univar object.}
\usage{
\method{plot}{hdensity_univar}(x, main = "Hermite PDF", xlab = "x", ylab = "Density", ...)
}
\arguments{
\item{x}{A hdensity_univar object.}

\item{main}{A string, title for plot.}

\item{xlab}{A string, x label for plot.}

\item{ylab}{A string, y label for plot.}

\item{...}{Additional parameters for plotting.}
}
\description{
Plots the hdensity_univar object as output by the density function when 
evaluated on a hermite_estimator_univar object.
}
