% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{print.hcdf_bivar}
\alias{print.hcdf_bivar}
\title{Prints the hcdf_bivar object as output by the hcdf function when evaluated
on a hermite_estimator_bivar object.}
\usage{
\method{print}{hcdf_bivar}(x, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{x}{A hcdf_bivar object.}

\item{digits}{A numeric value. Number of digits to round to.}

\item{...}{Additional parameters for printing.}
}
\description{
Prints the hcdf_bivar object as output by the hcdf function when evaluated
on a hermite_estimator_bivar object.
}
