% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{print.hcdf_univar}
\alias{print.hcdf_univar}
\title{Prints the hcdf_univar object as output by the hcdf function when evaluated
on a hermite_estimator_univar object.}
\usage{
\method{print}{hcdf_univar}(x, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{x}{A hcdf_univar object.}

\item{digits}{A numeric value. Number of digits to round to.}

\item{...}{Unused}
}
\description{
Mirrors the print method of the stats::ecdf function
}
