% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{print.hdensity_bivar}
\alias{print.hdensity_bivar}
\title{Prints the hdensity_bivar object as output by the density function when
evaluated on a hermite_estimator_bivar object.}
\usage{
\method{print}{hdensity_bivar}(x, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{x}{A hdensity_bivar object.}

\item{digits}{A numeric value. Number of digits to round to.}

\item{...}{Additional parameters for printing.}
}
\description{
Prints the hdensity_bivar object as output by the density function when
evaluated on a hermite_estimator_bivar object.
}
