% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{print.hdensity_univar}
\alias{print.hdensity_univar}
\title{Prints the hdensity_univar object as output by the density function when
evaluated on a hermite_estimator_univar object.}
\usage{
\method{print}{hdensity_univar}(x, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{x}{A hdensity_univar object.}

\item{digits}{A numeric value. Number of digits to round to.}

\item{...}{Unused}
}
\description{
Prints the hdensity_univar object as output by the density function when
evaluated on a hermite_estimator_univar object.
}
