% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_bivar.R
\name{print.hermite_estimator_bivar}
\alias{print.hermite_estimator_bivar}
\title{Prints bivariate hermite_estimator object.}
\usage{
\method{print}{hermite_estimator_bivar}(x, ...)
}
\arguments{
\item{x}{A hermite_estimator_bivar object.}

\item{...}{Other arguments passed on to methods used in printing.}
}
\description{
Prints bivariate hermite_estimator object.
}
