% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{print.hermite_estimator_univar}
\alias{print.hermite_estimator_univar}
\title{Prints univariate hermite_estimator object.}
\usage{
\method{print}{hermite_estimator_univar}(x, ...)
}
\arguments{
\item{x}{A hermite_estimator_univar object.}

\item{...}{Other arguments passed on to methods used in printing.}
}
\description{
Prints univariate hermite_estimator object.
}
\examples{
hermite_est <- hermite_estimator_univar(N = 10, standardize = TRUE)
print(hermite_est)
}
