% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{quantile.hermite_estimator_univar}
\alias{quantile.hermite_estimator_univar}
\title{Estimates the quantiles at a vector of probability values}
\usage{
\method{quantile}{hermite_estimator_univar}(x, probs = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{A hermite_estimator_univar object.}

\item{probs}{A numeric vector. A vector of probability values.}

\item{...}{Optional additional arguments to the quant function namely 
algorithm and accelerate_series.}
}
\value{
A numeric vector. The vector of quantile values associated with the
probabilities probs.
}
\description{
This generic method is a convenience wrapper around the quant method
}
