% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{quantile_helper}
\alias{quantile_helper}
\title{Estimates the quantile at a single probability value}
\usage{
quantile_helper(this, p)
}
\arguments{
\item{this}{A hermite_estimator object.}

\item{p}{A numeric value. The probability at which to calculate the
associated quantile.}
}
\value{
A numeric value. The value of the quantile associated with p.
}
\description{
This helper method is intended for internal use by the hermite_estimator
class.
}
