% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{standardize_value}
\alias{standardize_value}
\title{Standardize a vector of observations x}
\usage{
standardize_value(this, x)
}
\arguments{
\item{this}{A hermite_estimator object.}

\item{x}{A numeric vector of observations.}
}
\value{
An object of class hermite_estimator.
}
\description{
This helper method standardizes the observations x using the online mean and
online standard deviation contained in the hermite_estimator object (this).
}
