% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{standardizeInputsEW}
\alias{standardizeInputsEW}
\title{Standardizes the observation x and updates the online moment inputs}
\usage{
standardizeInputsEW(x, n_obs, lambda, current_mean, current_var)
}
\arguments{
\item{x}{A numeric value.}

\item{n_obs}{A numeric value. The number of observations.}

\item{lambda}{A numeric value.}

\item{current_mean}{A numeric value.}

\item{current_var}{A numeric value.}
}
\value{
A numeric vector. The first element is the updated mean. The
second element is the updated variance times n_obs. The third element is the
updated, standardized value of x.
}
\description{
The online moments are updated via exponential weighting.
}
\author{
Michael Stephanou <michael.stephanou@gmail.com>
}
