% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{summary.hcdf_bivar}
\alias{summary.hcdf_bivar}
\title{Summarizes the hcdf_bivar object as output by the hcdf function when 
evaluated on a hermite_estimator_bivar object.}
\usage{
\method{summary}{hcdf_bivar}(object, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{object}{A hcdf_bivar object.}

\item{digits}{A numeric value. Number of digits to round to.}

\item{...}{Additional parameters for printing.}
}
\description{
Summarizes the hcdf_bivar object as output by the hcdf function when 
evaluated on a hermite_estimator_bivar object.
}
