% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{summary.hcdf_univar}
\alias{summary.hcdf_univar}
\title{Summarizes the hcdf_univar object as output by the hcdf function when 
evaluated on a hermite_estimator_univar object.}
\usage{
\method{summary}{hcdf_univar}(object, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{object}{A hcdf_univar object.}

\item{digits}{A numeric value. Number of digits to round to.}

\item{...}{Unused.}
}
\description{
Summarizes the hcdf_univar object as output by the hcdf function when 
evaluated on a hermite_estimator_univar object.
}
