% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_bivar.R
\name{update_batch.hermite_estimator_bivar}
\alias{update_batch.hermite_estimator_bivar}
\title{Updates the Hermite series based estimator with a batch of data}
\usage{
\method{update_batch}{hermite_estimator_bivar}(this, x)
}
\arguments{
\item{this}{A hermite_estimator_bivar object.}

\item{x}{A numeric matrix. A matrix of bivariate observations to be 
incorporated into the estimator. Each row corresponds to a single bivariate
observation.}
}
\value{
An object of class hermite_estimator_bivar.
}
\description{
This method can be applied in one-pass batch estimation settings. This
method cannot be used with an exponentially weighted estimator.
}
\examples{
hermite_estimator <- hermite_estimator_bivar(N = 10, standardize = TRUE)
hermite_estimator <- update_batch(hermite_estimator, x = matrix(c(1, 2, 3, 4,
 5, 6),nrow=3, ncol=2, byrow = TRUE))
}
