% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{update_batch.hermite_estimator_univar}
\alias{update_batch.hermite_estimator_univar}
\title{Updates the Hermite series based estimator with a batch of data}
\usage{
\method{update_batch}{hermite_estimator_univar}(this, x)
}
\arguments{
\item{this}{A hermite_estimator_univar object.}

\item{x}{A numeric vector. A vector of observations to be incorporated
into the estimator.}
}
\value{
An object of class hermite_estimator_univar.
}
\description{
This method can be applied in one-pass batch estimation settings. This
method cannot be used with an exponentially weighted estimator.
}
\examples{
hermite_est <- hermite_estimator_univar(N = 10, standardize = TRUE)
hermite_est <- update_batch(hermite_est, x = c(1, 2))
}
