% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{update_batch}
\alias{update_batch}
\title{Updates the Hermite series based estimator with a batch of data}
\usage{
update_batch(this, x)
}
\arguments{
\item{this}{A hermite_estimator object.}

\item{x}{A numeric vector. A vector of observations to be incorporated
into the estimator.}
}
\value{
An object of class hermite_estimator.
}
\description{
This method can be applied in one-pass batch estimation settings. This
method cannot be used with an exponentially weighted estimator.
}
\examples{
hermite_estimator <- hermite_estimator(N = 10, standardize = TRUE)
hermite_estimator <- update_batch(hermite_estimator, x = c(1, 2))
}
