% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{update_sequential}
\alias{update_sequential}
\title{Updates the Hermite series based estimator sequentially}
\usage{
update_sequential(this, x)
}
\arguments{
\item{this}{A hermite_estimator object.}

\item{x}{A numeric value. An observation to be incorporated into the
estimator.}
}
\value{
An object of class hermite_estimator.
}
\description{
This method can be applied in sequential estimation settings.
}
\examples{
hermite_estimator <- hermite_estimator(N = 10, standardize = TRUE)
hermite_estimator <- update_sequential(hermite_estimator, x = 2)
}
