% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator.R
\name{combine_hermite}
\alias{combine_hermite}
\title{Combines a list of Hermite estimators}
\usage{
combine_hermite(hermite_estimators)
}
\arguments{
\item{hermite_estimators}{A list of hermite_estimator objects.}
}
\value{
An object of class hermite_estimator.
}
\description{
This method allows a list of Hermite based estimators of class 
hermite_estimator to be consistently combined.
}
\details{
Note that the N and normalize arguments must be the same for all estimators 
in order to combine them. In addition, note that exponentially weighted 
estimators cannot be combined.
}
