% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hermite_normalization}
\alias{hermite_normalization}
\title{Outputs Hermite normalization factors}
\usage{
hermite_normalization(N)
}
\arguments{
\item{N}{An integer number.}
}
\value{
A numeric vector of length N+1
}
\description{
The method returns numeric normalization factors that, when multiplied by 
the physicist Hermite polynomials times a Gaussian factor i.e.
\eqn{\exp{x^2/2}H_k(x)}, yields orthonormal 
Hermite functions \eqn{h_k(x)} for \eqn{k=0,\dots,N}.
}
\author{
Michael Stephanou <michael.stephanou@gmail.com>
}
