% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_bivar.R
\name{kendall.hermite_estimator_bivar}
\alias{kendall.hermite_estimator_bivar}
\title{Estimates the Kendall rank correlation coefficient}
\usage{
\method{kendall}{hermite_estimator_bivar}(h_est_obj, clipped = FALSE)
}
\arguments{
\item{h_est_obj}{A hermite_estimator_bivar object.}

\item{clipped}{A boolean value. Indicates whether to clip the Kendall rank 
correlation estimates to lie between -1 and 1.}
}
\value{
A numeric value.
}
\description{
This method calculates the Kendall rank correlation coefficient value
using the hermite_estimator_bivar object (h_est_obj).
}
\details{
The object must be updated with observations prior to the use of this method.
}
