% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_bivar.R
\name{update_sequential.hermite_estimator_bivar}
\alias{update_sequential.hermite_estimator_bivar}
\title{Updates the Hermite series based estimator sequentially}
\usage{
\method{update_sequential}{hermite_estimator_bivar}(h_est_obj, x)
}
\arguments{
\item{h_est_obj}{A hermite_estimator_bivar object.}

\item{x}{A numeric vector of length 2. A bivariate observation to be 
incorporated into the estimator.}
}
\value{
An object of class hermite_estimator_bivar.
}
\description{
This method can be applied in sequential estimation settings.
}
