% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermite_estimator_univar.R
\name{update_sequential.hermite_estimator_univar}
\alias{update_sequential.hermite_estimator_univar}
\title{Updates the Hermite series based estimator sequentially}
\usage{
\method{update_sequential}{hermite_estimator_univar}(this, x)
}
\arguments{
\item{this}{A hermite_estimator_univar object.}

\item{x}{A numeric value. An observation to be incorporated into the
estimator.}
}
\value{
An object of class hermite_estimator_univar.
}
\description{
This method can be applied in sequential estimation settings.
}
\examples{
hermite_est <- hermite_estimator_univar(N = 10, standardize = TRUE)
hermite_est <- update_sequential(hermite_est, x = 2)
}
